Nodes are now using the "component paradigm".
- Each node contains N components into Node::components member.
- For now components are used only to attach a view to a node but as I wanted Nodable to be extendable I do it now.
- You can set a component view by calling void myNode.addComponent("view", new MyViewDerivedClass(&node))
- You can ask if a node has a component by calling bool myNode.hasComponent("view");
- You can get the view by calling  Node* Node::getComponent("view"); and casting the result to MyViewDerivedClass* for instance.

+++ sources/Node_Container.h
@@ -21,7 +21,7 @@ namespace Nodable{
 		void                      addNode                   (Node* /*Node to add to this context*/);
 		void                      destroyNode               (Node*);
 		Node_Variable*            createNodeVariable        (const char* /*name*/ = "");
-		Node_Variable*            createNodeNumber          (int /*value*/ = 0);
+		Node_Variable*            createNodeNumber          (double /*value*/ = 0);
 		Node_Variable*            createNodeNumber          (const char* /*value*/);
 		Node_Variable*            createNodeString          (const char* /*value*/);
 		Node_BinaryOperation*     createNodeBinaryOperation (std::string /*_operator*/, Node_Variable* /*_leftInput*/, Node_Variable*/*_rightInput*/, Node_Variable*/*output*/);
